Split config into separate package
