+++ handlers/digital-subscription-expiry/src/test/scala/com/gu/digitalSubscriptionExpiry/zuora/GetAccountSummaryEffectsTest.scala
@@ -3,10 +3,10 @@ import java.io
 
 import com.gu.digitalSubscriptionExpiry.Handler.StepsConfig
 import com.gu.digitalSubscriptionExpiry.zuora.GetAccountSummary.{AccountId, AccountSummaryResult}
-import com.gu.effects.{ConfigLoad, RawEffects}
+import com.gu.effects.{S3ConfigLoad, RawEffects}
 import com.gu.test.EffectsTest
 import com.gu.util.zuora.ZuoraRestRequestMaker
-import com.gu.util.{Config, Stage}
+import com.gu.util.config.{LoadConfig, Stage}
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.syntax.std.either._
 import scalaz.{\/, \/-}
@@ -16,8 +16,8 @@ class GetAccountSummaryEffectsTest extends FlatSpec with Matchers {
     val testAccountId = AccountId("2c92c0f86078c4d4016079e1402d6536")
 
     val actual: \/[io.Serializable, AccountSummaryResult] = for {
-      configAttempt <- ConfigLoad.load(Stage("DEV")).toEither.disjunction
-      config <- Config.parseConfig[StepsConfig](configAttempt)
+      configAttempt <- S3ConfigLoad.load(Stage("DEV")).toEither.disjunction
+      config <- LoadConfig.parseConfig[StepsConfig](configAttempt)
       zuoraRequests = ZuoraRestRequestMaker(RawEffects.response, config.stepsConfig.zuoraRestConfig)
       subscription <- GetAccountSummary(zuoraRequests)(testAccountId)
     } yield {
