+++ handlers/identity-backfill/src/test/scala/com/gu/identity/GetByEmailEffectsTest.scala
@@ -1,10 +1,10 @@
 package com.gu.identity
 
-import com.gu.effects.{ConfigLoad, RawEffects}
+import com.gu.effects.{S3ConfigLoad, RawEffects}
 import com.gu.identityBackfill.Handler.StepsConfig
 import com.gu.identityBackfill.Types.{EmailAddress, IdentityId}
 import com.gu.test.EffectsTest
-import com.gu.util.{Config, Stage}
+import com.gu.util.config.{LoadConfig, Stage}
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.\/-
 import scalaz.syntax.std.either._
@@ -15,8 +15,8 @@ class GetByEmailEffectsTest extends FlatSpec with Matchers {
   it should "successfull run the health check using the local code against real backend" taggedAs EffectsTest in {
 
     val actual = for {
-      configAttempt <- ConfigLoad.load(Stage("DEV")).toEither.disjunction
-      config <- Config.parseConfig[StepsConfig](configAttempt)
+      configAttempt <- S3ConfigLoad.load(Stage("DEV")).toEither.disjunction
+      config <- LoadConfig.parseConfig[StepsConfig](configAttempt)
       identityId <- GetByEmail(RawEffects.response, config.stepsConfig.identityConfig)(EmailAddress("john.duffell@guardian.co.uk"))
     } yield {
       identityId
