+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/salesforce/DevSFEffects.scala
@@ -2,7 +2,7 @@ package com.gu.identityBackfill.salesforce
 
 import com.gu.effects.RawEffects
 import com.gu.identityBackfill.Handler.StepsConfig
-import com.gu.util.{Config, Stage}
+import com.gu.util.config.{LoadConfig, Stage}
 import okhttp3.{Request, Response}
 import scalaz.syntax.std.either._
 
@@ -10,7 +10,7 @@ object DevSFEffects {
   def apply(effects: RawEffects, response: Request => Response) = {
     for {
       configAttempt <- effects.s3Load(Stage("DEV")).toEither.disjunction
-      config <- Config.parseConfig[StepsConfig](configAttempt)
+      config <- LoadConfig.parseConfig[StepsConfig](configAttempt)
       auth <- SalesforceAuthenticate(response, config.stepsConfig.sfConfig)
     } yield auth
   }
