+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/salesforce/SalesforceAuthenticateEffectsTest.scala
@@ -1,10 +1,10 @@
 package com.gu.identityBackfill.salesforce
 
-import com.gu.effects.{ConfigLoad, RawEffects}
+import com.gu.effects.{RawEffects, S3ConfigLoad}
 import com.gu.identityBackfill.Handler.StepsConfig
 import com.gu.identityBackfill.salesforce.SalesforceAuthenticate.SalesforceAuth
 import com.gu.test.EffectsTest
-import com.gu.util.{Config, Stage}
+import com.gu.util.config.{LoadConfig, Stage}
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.\/-
 import scalaz.syntax.std.either._
@@ -14,8 +14,8 @@ class SalesforceAuthenticateEffectsTest extends FlatSpec with Matchers {
   it should "get auth SF correctly" taggedAs EffectsTest in {
 
     val actual = for {
-      configAttempt <- ConfigLoad.load(Stage("DEV")).toEither.disjunction
-      config <- Config.parseConfig[StepsConfig](configAttempt)
+      configAttempt <- S3ConfigLoad.load(Stage("DEV")).toEither.disjunction
+      config <- LoadConfig.parseConfig[StepsConfig](configAttempt)
       identityId <- SalesforceAuthenticate.doAuth(RawEffects.response, config.stepsConfig.sfConfig)
     } yield {
       identityId
