+++ lib/effects/src/test/scala/com/gu/effects/TestingRawEffects.scala
@@ -5,7 +5,8 @@ import java.io.ByteArrayInputStream
 import com.amazonaws.AmazonServiceException
 import com.amazonaws.services.s3.model.{ObjectMetadata, PutObjectRequest, PutObjectResult}
 import com.gu.effects.TestingRawEffects._
-import com.gu.util.{Logging, Stage}
+import com.gu.util.Logging
+import com.gu.util.config.Stage
 import okhttp3._
 import okhttp3.internal.Util.UTF_8
 import okio.Buffer
