+++ lib/handler/src/test/scala/com/gu/util/AuthTest.scala
@@ -1,7 +1,8 @@
 package com.gu.util
 
-import com.gu.util.Auth._
+import com.gu.util.apigateway.Auth._
 import com.gu.util.apigateway.{ApiGatewayRequest, RequestAuth}
+import com.gu.util.config.TrustedApiConfig
 import org.scalatest.FlatSpec
 import play.api.libs.json.{JsError, JsSuccess, JsValue, Json}
