+++ src/main/scala/com/gu/stripeCustomerSourceUpdated/StripeSignatureChecker.scala
@@ -1,7 +1,8 @@
 package com.gu.stripeCustomerSourceUpdated
 
 import com.gu.util.apigateway.StripeAccount
-import com.gu.util.{Logging, StripeConfig, StripeSecretKey}
+import com.gu.util.config.{StripeConfig, StripeSecretKey}
+import com.gu.util.Logging
 import com.stripe.exception.SignatureVerificationException
 import com.stripe.net.Webhook.Signature
