+++ src/test/scala/manualTest/EmailClientSystemTest.scala
@@ -1,13 +1,13 @@
 package manualTest
 
-import com.gu.effects.{ConfigLoad, RawEffects}
+import com.gu.effects.{S3ConfigLoad, RawEffects}
 import com.gu.stripeCustomerSourceUpdated.SourceUpdatedSteps.StepsConfig
-import com.gu.util.ETConfig.ETSendIds
+import com.gu.util.config.ETConfig.ETSendIds
 import com.gu.util.exacttarget._
 import com.gu.util.reader.Types._
-import com.gu.util.{Config, Logging, Stage}
+import com.gu.util.Logging
+import com.gu.util.config.{LoadConfig, Stage}
 import scalaz.syntax.std.either._
-
 import scala.util.Random
 
 // run this to send a one off email to yourself.  the email will take a few mins to arrive, but it proves the ET logic works
@@ -48,8 +48,8 @@ object EmailClientSystemTest extends App with Logging {
     )
 
   for {
-    configAttempt <- ConfigLoad.load(Stage("DEV")).toEither.disjunction.withLogging("fromFile")
-    config <- Config.parseConfig[StepsConfig](configAttempt)
+    configAttempt <- S3ConfigLoad.load(Stage("DEV")).toEither.disjunction.withLogging("fromFile")
+    config <- LoadConfig.parseConfig[StepsConfig](configAttempt)
     send = EmailSendSteps(ETClient.sendEmail(RawEffects.response, config.etConfig), FilterEmail(Stage("CODE")))_
     etSendIds = config.etConfig.etSendIDs
   } yield Seq(
